Se creo todo lo de Student, StudentConsumption, Inventory School
Además se realizaron algunas actualizaciones generales

+++ GestionPae/src/main/java/com/example/GestionPae/model/Student.java
@@ -16,7 +16,7 @@ public class Student {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int idStudent;
+    private Long idStudent;
 
     @ManyToOne
     @JoinColumn(name = "id_user", nullable = false)
